pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.20.0",
    "google-cloud-translate>=3.22.0",
    "google-cloud-vision>=3.11.0",
    "google-genai>=1.47.0",
    "googlemaps>=4.10.0",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pypdf2>=3.0.1",
    "streamlit>=1.51.0",
    "streamlit-folium>=0.25.3",
]
